+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasetsop.py
@@ -113,7 +113,7 @@ class ListDatasetsData(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-06-01T13:48:56.269Z")
+    updated: Optional[datetime] = parse_datetime("2026-06-01T14:43:52.579Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")