+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KT1Q4DVRWW35DZRWFHV8A76G"
+        "01KT1T90XAMN29NDXVN72JF8FV"
     )
     r"""The unique identifier of the data source"""