+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KT1Q4DVT4RD40NAWND08EJV3"
+        "01KT1T90XC39T1CQ2V9E7Z207X"
     )
     r"""The unique identifier of the data source"""