+++ packages/orq-rc/src/orq_ai_sdk/models/streamrunagentop.py
@@ -1903,7 +1903,7 @@ class AgentToolInputRunTools(BaseModel):
         StreamRunAgentAgentToolInputRunAgentsSchema, pydantic.Field(alias="schema")
     ]
 
-    id: Optional[str] = "01KT1Q4DJ4Y0FFJDTPJTW9CZMT"
+    id: Optional[str] = "01KT1T90KK80N02YY5SRWE59EA"
 
     description: Optional[str] = None