+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KT1Q4DVYT7XZX34J49Z9SPT9"
+        "01KT1T90XZDDQZY3HDVQS4H4RY"
     )
     r"""The unique identifier of the data source"""