+++ packages/orq-rc/RELEASES.md
@@ -8339,3 +8339,13 @@ Based on:
 - [python v4.10.0-rc.64] packages/orq-rc
 ### Releases
 - [PyPI v4.10.0-rc.64] https://pypi.org/project/orq-ai-sdk/4.10.0-rc.64 - packages/orq-rc
+
+## 2026-06-01 14:46:29
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.769.1 (2.892.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.10.0-rc.65] packages/orq-rc
+### Releases
+- [PyPI v4.10.0-rc.65] https://pypi.org/project/orq-ai-sdk/4.10.0-rc.65 - packages/orq-rc